requirements.txt
azure-messaging-webpubsubservice==1.0.1
python-dotenv==1.0.1
watchdog==4.0.0
websockets==12.0
